requirements.txt
numpy<2
opencv-python
Pillow
pytesseract
transformers<5
torch==2.2.2
openai
python-dotenv
datasets
